# docs/swagger.yaml
...
info:
  contact: {}
  description: Контент сервис - сервис хранения контента (баннеры, тесты, истории
    и т.д.) для дальнейшего отображения его на клиенте пользователя.
  title: Content Service API
  version: "1.0"
paths:
...
  /scenario/{user_id}:
    get:
      description: Метод получение всего контента, который доступен пользователю по
        его сценарию.
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение всего контента по сценарию пользователя.
      tags:
      - User
  /scenario/block/{user_id}/{block_id}:
    get:
      description: Метод получения элемента контента из блока по id, названию или
        типу контента.
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: integer
      - description: id блока
        in: path
        name: block_id
        required: true
        type: integer
      - description: id контента
        in: query
        name: content_id
        type: integer
      - description: название элемента контента
        in: query
        name: name
        type: string
      - description: тип контента (1 - баннер, 2 - истории, 3 - тесты)
        in: query
        name: content_type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение элемента контента из блока по id, названию или типу контента.
      tags:
      - User
  /scenario/block/{user_id}/list/{block_id}:
    get:
      description: Метод получения списка id и названий элементов контента для блока.
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: integer
      - description: id блока
        in: path
        name: block_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение списка id и названий элементов контента для блока.
      tags:
      - User
  /scenario/blocks/{user_id}:
    get:
      description: Метод получение блока контента определенного типа по id и/или названию.
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: integer
      - description: id блока
        in: query
        name: block_id
        type: integer
      - description: название блока
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение блока контента по id и/или названию для пользователя.
      tags:
      - User
  /scenario/list/{user_id}:
    get:
      description: Метод получения списка id и названий блоков, доступных пользователю
        по сценарию.
      parameters:
      - description: id пользователя
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение списка id и названий блоков, доступных пользователю по сценарию.
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"